Fix compilation warnings on GCC 4.0; fix scons 0.96.91 compatibility

+++ trunk/grftool/gtk/sprite-viewer.cpp
@@ -7,7 +7,7 @@ animate (SpriteViewer *viewer)
 		viewer->frames[viewer->currentFrame]);
 
 	viewer->currentFrame++;
-	if (viewer->currentFrame >= viewer->frames.size ())
+	if (viewer->currentFrame >= (int) viewer->frames.size ())
 		viewer->currentFrame = 0;
 	return TRUE;
 }
